🗑️ zb: Deprecate Connection::peer_credentials
Since this method doesn't and can't cache the credentials, we deprecate
it in favor of `peer_creds` that does cache. The reason the former can't
cache is because it returns `Credentials` struct by value and
`Credentials` struct can't simply be cloned because it contains an FD
(PIDFD on Linux) and duplicating that behind user's back isn't something
we should do. We already do that for FD passing but it couldn't be
avoided. We could also change `peer_credentials` to return a reference
instead but that would be an API break.