➖ xmlgen: Allow the library part not depend on clap
By making clap an optional dependency and requiring it as a feature by
the binary, we ensure that the library part can be built w/o clap as
a dependency if default features are disabled.

Technically, this is a breaking change since
`cargo build --bin zbus-xmlgen --no-default-features` will fail to build
now but given that we had no features in the binary and this works for the
library part, I think it's hopefully not going to break any builds and if
it does, it's an easy fix.

Co-authored-by: @HigherOrderLogic